Added runner script to run datasets automatically
In this commit, I've added a runner.sh BASH scipt that compiles and runs the error correction code for each of the sequence files (in .txt format) in a given folder. The script takes in two parameters as input: the path to the error correction code and the path to the folder containing the dataset sequences. It runs the algorithm for each of the datasets, and stores the results obtained after each run in a separate "Results" folder. All output files are renamed appropriately in each run of the code to indicate which dataset they pertain to.