Removed other codebases from the repository
In this commit, I've removed the folders containing the original codebases for Lighter, KmerCo, and RobustBF from this repository. Those folders were nothing but clones of their original Github repositories, and are now redundant. These codebases have been cited in our final report as well as our README file, and anyone wishing to access them can refer to the links and visit the original Github repositories of the same.